unecessary big refactor, to fix ts err
