+++ src/react/utilsApp.ts
@@ -2,7 +2,6 @@ import { proxy, useSnapshot } from 'valtio'
 import { useEffect, useMemo } from 'react'
 import { useMedia } from 'react-use'
 import { activeModalStack, miscUiState } from '../globalState'
-import { currentScaling } from '../scaleInterface'
 
 export const watchedModalsFromHooks = proxy({
   value: new Set<string>()
@@ -37,19 +36,4 @@ export const useIsWidgetActive = (name: string) => {
 
 export const useIsSmallWidth = () => {
   return useMedia('(max-width: 550px)')
-}
-
-export const usePassesScaledDimensions = (minWidth: number | null = null, minHeight: number | null = null) => {
-  const { scale } = useSnapshot(currentScaling)
-  const conditions: string[] = []
-
-  if (minWidth !== null) {
-    conditions.push(`(min-width: ${minWidth * scale}px)`)
-  }
-  if (minHeight !== null) {
-    conditions.push(`(min-height: ${minHeight * scale}px)`)
-  }
-
-  const media = conditions.join(' and ') || 'all'
-  return useMedia(media)
 }
